refactor(menugroup): menugroup 패키지 별도로 생성
