+++ src/main/java/kitchenpos/menus/tobe/application/MenuGroupService.java
@@ -2,7 +2,7 @@ package kitchenpos.menus.tobe.application;
 
 import java.util.List;
 import kitchenpos.common.tobe.domain.DisplayedName;
-import kitchenpos.menus.tobe.domain.model.MenuGroup;
+import kitchenpos.menugroup.model.MenuGroup;
 import kitchenpos.menus.tobe.domain.repository.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
