+++ src/main/java/kitchenpos/menus/tobe/domain/repository/MenuGroupRepository.java
@@ -3,7 +3,7 @@ package kitchenpos.menus.tobe.domain.repository;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import kitchenpos.menus.tobe.domain.model.MenuGroup;
+import kitchenpos.menugroup.model.MenuGroup;
 
 public interface MenuGroupRepository {
     MenuGroup save(MenuGroup menuGroup);
