+++ src/main/java/kitchenpos/menus/tobe/infra/JpaMenuGroupRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.tobe.infra;
 
 import java.util.UUID;
-import kitchenpos.menus.tobe.domain.model.MenuGroup;
+import kitchenpos.menugroup.model.MenuGroup;
 import kitchenpos.menus.tobe.domain.repository.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
